Isolated userData to its own type and added file for query return types
