+++ src/types/loginTypes.ts
@@ -1,3 +1,5 @@
+import { UserData } from "./userTypes";
+
 export interface LoginData
 {
 	username: string,
@@ -8,25 +10,5 @@ export interface LoginReturnPackage
 {
 	success: boolean,
 	error: string,
-	userID: number,
-	username: string,
-	firstname: string,
-	lastname: string,
-	address: string,
-	lastUpdate: number,
-	balance: number,
-	spent: number
-};
-
-export interface UserData
-{
-	ID: number,
-	username: string,
-	password: string,
-	firstname: string,
-	lastname: string,
-	address: string,
-	lastUpdate: number,
-	balance: number,
-	spent: number
+	userData: UserData
 };
