+++ src/types/queryReturnTypes.ts
@@ -1,0 +1,12 @@
+export interface UsersQueryReturn
+{
+	id: number,
+	username: string,
+	password: string,
+	firstname: string,
+	lastname: string,
+	address: string,
+	lastUpdate: number,
+	balance: number,
+	spent: number
+};
