+++ src/types/userTypes.ts
@@ -1,0 +1,11 @@
+export interface UserData
+{
+	userID: number,
+	username: string,
+	firstname: string,
+	lastname: string,
+	address: string,
+	lastUpdate: number,
+	balance: number,
+	spent: number
+}
